refactor: extract AbstractStaticResourceHandlerFactory
Extracts `AbstractStaticResourceHandlerFactory` from the `StaticMappedResourceHandlerFactory` and `StaticResourceHandlerFactory`, defining the `configureMiddleware()` method, and declaring the `__invoke()` method.
